feat(application-layout): refactor application layout
