+++ apps/frontend/src/components/layouts/auditor/rs_layout.tsx
@@ -1,32 +1,27 @@
 import React from "react";
-import DashBoardPageLayout from "@components/layouts/dashboard_page_layout";
+import DashBoardPageLayout from "@components/layouts/auditor/dashboard/dashboard_page_layout";
 import { useRouter } from "next/router";
 
-interface FinancialsLayoutProps {
+interface researchLayoutProps {
   children: React.ReactNode;
   showexport: boolean;
 }
-export default function FinancialsLayout({
+export default function ResearchLayout({
   children,
   showexport,
-}: FinancialsLayoutProps) {
+}: researchLayoutProps) {
   const router = useRouter();
   const applicationID = router.query.applicationId;
   const SortbarData = [
     {
       id: 1,
-      title: "Profit & Loss",
-      path: `/applications/${applicationID}/financial/profit`,
+      title: "Research",
+      path: `/applications/${applicationID}/research/research`,
     },
     {
       id: 2,
-      title: "Balance Sheet",
-      path: `/applications/${applicationID}/financial/balance`,
-    },
-    {
-      id: 3,
-      title: "Cash Flow",
-      path: `/applications/${applicationID}/financial/cash`,
+      title: "Sentiment",
+      path: `/applications/${applicationID}/research/sentiments`,
     },
   ];
   return (
