+++ apps/frontend/src/pages/applications/[applicationId]/rating/qualitative/index.tsx
@@ -1,4 +1,4 @@
-import CRSLayout from "@components/layouts/crs_layout";
+import CRSLayout from "@components/layouts/auditor/crs_layout";
 import { useRouter } from "next/router";
 import React from "react";
