+++ apps/frontend/src/pages/applications/[applicationId]/research/research/index.tsx
@@ -1,5 +1,5 @@
 import { PortalIcons } from "@components/icons";
-import ResearchLayout from "@components/layouts/rs_layout";
+import ResearchLayout from "@components/layouts/auditor/rs_layout";
 import { Button, ResearchTable } from "taral-ui";
 import { useState } from "react";
 import { useForm } from "react-hook-form";
