+++ apps/frontend/src/pages/applications/[applicationId]/status/audit/index.tsx
@@ -1,4 +1,4 @@
-import StatusLayout from "@components/layouts/status_layout";
+import StatusLayout from "@components/layouts/auditor/status_layout";
 import { AuditTable } from "taral-ui";
 
 const TableData = [
