+++ apps/frontend/src/pages/applications/[applicationId]/status/customer_comms/index.tsx
@@ -1,5 +1,5 @@
 import { Activity } from "taral-ui";
-import StatusLayout from "@components/layouts/status_layout";
+import StatusLayout from "@components/layouts/auditor/status_layout";
 export default function CustomerComms() {
   return (
     <StatusLayout showexport={false}>
